citation to citation2 (fix neg sample bug)
