+++ examples/linkproppred/citation2/cluster_gcn.py
@@ -163,7 +163,7 @@ def test(model, predictor, data, split_edge, evaluator, batch_size, device):
 
 
 def main():
-    parser = argparse.ArgumentParser(description='OGBL-Citation (Cluster-GCN)')
+    parser = argparse.ArgumentParser(description='OGBL-Citation2 (Cluster-GCN)')
     parser.add_argument('--device', type=int, default=0)
     parser.add_argument('--log_steps', type=int, default=1)
     parser.add_argument('--num_partitions', type=int, default=15000)
@@ -182,7 +182,7 @@ def main():
     device = f'cuda:{args.device}' if torch.cuda.is_available() else 'cpu'
     device = torch.device(device)
 
-    dataset = PygLinkPropPredDataset(name='ogbl-citation')
+    dataset = PygLinkPropPredDataset(name='ogbl-citation2')
     split_edge = dataset.get_edge_split()
     data = dataset[0]
     data.edge_index = to_undirected(data.edge_index, data.num_nodes)
@@ -207,7 +207,7 @@ def main():
     predictor = LinkPredictor(args.hidden_channels, args.hidden_channels, 1,
                               args.num_layers, args.dropout).to(device)
 
-    evaluator = Evaluator(name='ogbl-citation')
+    evaluator = Evaluator(name='ogbl-citation2')
     logger = Logger(args.runs, args)
 
     for run in range(args.runs):
