+++ examples/linkproppred/citation2/gnn.py
@@ -1,14 +1,12 @@
 import argparse
 
 import torch
-import numpy as np
+from torch.nn import Parameter
 import torch.nn.functional as F
 from torch.utils.data import DataLoader
 
-from torch_sparse import SparseTensor
-from torch_geometric.data import GraphSAINTRandomWalkSampler
-from torch_geometric.nn import GCNConv
-from torch_geometric.utils import to_undirected
+import torch_geometric.transforms as T
+from torch_geometric.nn import GCNConv, SAGEConv
 
 from ogb.linkproppred import PygLinkPropPredDataset, Evaluator
 
@@ -21,10 +19,13 @@ class GCN(torch.nn.Module):
         super(GCN, self).__init__()
 
         self.convs = torch.nn.ModuleList()
-        self.convs.append(GCNConv(in_channels, hidden_channels))
+        self.convs.append(
+            GCNConv(in_channels, hidden_channels, normalize=False))
         for _ in range(num_layers - 2):
-            self.convs.append(GCNConv(hidden_channels, hidden_channels))
-        self.convs.append(GCNConv(hidden_channels, out_channels))
+            self.convs.append(
+                GCNConv(hidden_channels, hidden_channels, normalize=False))
+        self.convs.append(
+            GCNConv(hidden_channels, out_channels, normalize=False))
 
         self.dropout = dropout
 
@@ -32,24 +33,38 @@ class GCN(torch.nn.Module):
         for conv in self.convs:
             conv.reset_parameters()
 
-    def forward(self, x, edge_index):
+    def forward(self, x, adj_t):
         for conv in self.convs[:-1]:
-            x = conv(x, edge_index)
+            x = conv(x, adj_t)
             x = F.relu(x)
             x = F.dropout(x, p=self.dropout, training=self.training)
-        x = self.convs[-1](x, edge_index)
+        x = self.convs[-1](x, adj_t)
         return x
 
 
-class GCNInference(torch.nn.Module):
-    def __init__(self, weights):
-        super(GCNInference, self).__init__()
-        self.weights = weights
+class SAGE(torch.nn.Module):
+    def __init__(self, in_channels, hidden_channels, out_channels, num_layers,
+                 dropout):
+        super(SAGE, self).__init__()
 
-    def forward(self, x, adj):
-        for i, (weight, bias) in enumerate(self.weights):
-            x = adj @ x @ weight + bias
-            x = np.clip(x, 0, None) if i < len(self.weights) - 1 else x
+        self.convs = torch.nn.ModuleList()
+        self.convs.append(SAGEConv(in_channels, hidden_channels))
+        for _ in range(num_layers - 2):
+            self.convs.append(SAGEConv(hidden_channels, hidden_channels))
+        self.convs.append(SAGEConv(hidden_channels, out_channels))
+
+        self.dropout = dropout
+
+    def reset_parameters(self):
+        for conv in self.convs:
+            conv.reset_parameters()
+
+    def forward(self, x, adj_t):
+        for conv in self.convs[:-1]:
+            x = conv(x, adj_t)
+            x = F.relu(x)
+            x = F.dropout(x, p=self.dropout, training=self.training)
+        x = self.convs[-1](x, adj_t)
         return x
 
 
@@ -80,23 +95,28 @@ class LinkPredictor(torch.nn.Module):
         return torch.sigmoid(x)
 
 
-def train(model, predictor, loader, optimizer, device):
+def train(model, predictor, data, split_edge, optimizer, batch_size):
     model.train()
+    predictor.train()
+
+    source_edge = split_edge['train']['source_node'].to(data.x.device)
+    target_edge = split_edge['train']['target_node'].to(data.x.device)
 
     total_loss = total_examples = 0
-    for data in loader:
-        data = data.to(device)
+    for perm in DataLoader(range(source_edge.size(0)), batch_size,
+                           shuffle=True):
         optimizer.zero_grad()
 
-        h = model(data.x, data.edge_index)
+        h = model(data.x, data.adj_t)
 
-        src, dst = data.edge_index
+        src, dst = source_edge[perm], target_edge[perm]
+
         pos_out = predictor(h[src], h[dst])
         pos_loss = -torch.log(pos_out + 1e-15).mean()
 
         # Just do some trivial random sampling.
-        dst_neg = torch.randint(0, data.x.size(0), src.size(),
-                                dtype=torch.long, device=device)
+        dst_neg = torch.randint(0, data.num_nodes, src.size(),
+                                dtype=torch.long, device=h.device)
         neg_out = predictor(h[src], h[dst_neg])
         neg_loss = -torch.log(1 - neg_out + 1e-15).mean()
 
@@ -104,7 +124,7 @@ def train(model, predictor, loader, optimizer, device):
         loss.backward()
         optimizer.step()
 
-        num_examples = src.size(0)
+        num_examples = pos_out.size(0)
         total_loss += loss.item() * num_examples
         total_examples += num_examples
 
@@ -112,29 +132,15 @@ def train(model, predictor, loader, optimizer, device):
 
 
 @torch.no_grad()
-def test(model, predictor, data, split_edge, evaluator, batch_size, device):
+def test(model, predictor, data, split_edge, evaluator, batch_size):
     predictor.eval()
-    print('Evaluating full-batch GNN on CPU...')
 
-    weights = [(conv.weight.cpu().detach().numpy(),
-                conv.bias.cpu().detach().numpy()) for conv in model.convs]
-    model = GCNInference(weights)
-
-    x = data.x.numpy()
-    adj = SparseTensor(row=data.edge_index[0], col=data.edge_index[1])
-    adj = adj.set_diag()
-    deg = adj.sum(dim=1)
-    deg_inv_sqrt = deg.pow(-0.5)
-    deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
-    adj = deg_inv_sqrt.view(-1, 1) * adj * deg_inv_sqrt.view(1, -1)
-    adj = adj.to_scipy(layout='csr')
-
-    h = torch.from_numpy(model(x, adj)).to(device)
+    h = model(data.x, data.adj_t)
 
     def test_split(split):
-        source = split_edge[split]['source_node'].to(device)
-        target = split_edge[split]['target_node'].to(device)
-        target_neg = split_edge[split]['target_node_neg'].to(device)
+        source = split_edge[split]['source_node'].to(h.device)
+        target = split_edge[split]['target_node'].to(h.device)
+        target_neg = split_edge[split]['target_node_neg'].to(h.device)
 
         pos_preds = []
         for perm in DataLoader(range(source.size(0)), batch_size):
@@ -163,18 +169,17 @@ def test(model, predictor, data, split_edge, evaluator, batch_size, device):
 
 
 def main():
-    parser = argparse.ArgumentParser(description='OGBL-Citation (GraphSAINT)')
+    parser = argparse.ArgumentParser(description='OGBL-Citation2 (GNN)')
     parser.add_argument('--device', type=int, default=0)
     parser.add_argument('--log_steps', type=int, default=1)
+    parser.add_argument('--use_sage', action='store_true')
     parser.add_argument('--num_layers', type=int, default=3)
     parser.add_argument('--hidden_channels', type=int, default=256)
-    parser.add_argument('--dropout', type=float, default=0.0)
-    parser.add_argument('--batch_size', type=int, default=16 * 1024)
-    parser.add_argument('--walk_length', type=int, default=3)
-    parser.add_argument('--lr', type=float, default=0.001)
-    parser.add_argument('--epochs', type=int, default=200)
-    parser.add_argument('--num_steps', type=int, default=100)
-    parser.add_argument('--eval_steps', type=int, default=10)
+    parser.add_argument('--dropout', type=float, default=0)
+    parser.add_argument('--batch_size', type=int, default=64 * 1024)
+    parser.add_argument('--lr', type=float, default=0.0005)
+    parser.add_argument('--epochs', type=int, default=50)
+    parser.add_argument('--eval_steps', type=int, default=1)
     parser.add_argument('--runs', type=int, default=10)
     args = parser.parse_args()
     print(args)
@@ -182,16 +187,13 @@ def main():
     device = f'cuda:{args.device}' if torch.cuda.is_available() else 'cpu'
     device = torch.device(device)
 
-    dataset = PygLinkPropPredDataset(name='ogbl-citation')
-    split_edge = dataset.get_edge_split()
+    dataset = PygLinkPropPredDataset(name='ogbl-citation2',
+                                     transform=T.ToSparseTensor())
     data = dataset[0]
-    data.edge_index = to_undirected(data.edge_index, data.num_nodes)
+    data.adj_t = data.adj_t.to_symmetric()
+    data = data.to(device)
 
-    loader = GraphSAINTRandomWalkSampler(data, batch_size=args.batch_size,
-                                         walk_length=args.walk_length,
-                                         num_steps=args.num_steps,
-                                         sample_coverage=0,
-                                         save_dir=dataset.processed_dir)
+    split_edge = dataset.get_edge_split()
 
     # We randomly pick some training samples that we want to evaluate on:
     torch.manual_seed(12345)
@@ -202,12 +204,27 @@ def main():
         'target_node_neg': split_edge['valid']['target_node_neg'],
     }
 
-    model = GCN(data.x.size(-1), args.hidden_channels, args.hidden_channels,
-                args.num_layers, args.dropout).to(device)
+    if args.use_sage:
+        model = SAGE(data.num_features, args.hidden_channels,
+                     args.hidden_channels, args.num_layers,
+                     args.dropout).to(device)
+    else:
+        model = GCN(data.num_features, args.hidden_channels,
+                    args.hidden_channels, args.num_layers,
+                    args.dropout).to(device)
+
+        # Pre-compute GCN normalization.
+        adj_t = data.adj_t.set_diag()
+        deg = adj_t.sum(dim=1).to(torch.float)
+        deg_inv_sqrt = deg.pow(-0.5)
+        deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
+        adj_t = deg_inv_sqrt.view(-1, 1) * adj_t * deg_inv_sqrt.view(1, -1)
+        data.adj_t = adj_t
+
     predictor = LinkPredictor(args.hidden_channels, args.hidden_channels, 1,
                               args.num_layers, args.dropout).to(device)
 
-    evaluator = Evaluator(name='ogbl-citation')
+    evaluator = Evaluator(name='ogbl-citation2')
     logger = Logger(args.runs, args)
 
     for run in range(args.runs):
@@ -216,24 +233,28 @@ def main():
         optimizer = torch.optim.Adam(
             list(model.parameters()) + list(predictor.parameters()),
             lr=args.lr)
+
         for epoch in range(1, 1 + args.epochs):
-            loss = train(model, predictor, loader, optimizer, device)
+            loss = train(model, predictor, data, split_edge, optimizer,
+                         args.batch_size)
             print(f'Run: {run + 1:02d}, Epoch: {epoch:02d}, Loss: {loss:.4f}')
 
-            if epoch > 49 and epoch % args.eval_steps == 0:
+            if epoch % args.eval_steps == 0:
                 result = test(model, predictor, data, split_edge, evaluator,
-                              batch_size=64 * 1024, device=device)
+                              args.batch_size)
                 logger.add_result(run, result)
 
-                train_mrr, valid_mrr, test_mrr = result
-                print(f'Run: {run + 1:02d}, '
-                      f'Epoch: {epoch:02d}, '
-                      f'Loss: {loss:.4f}, '
-                      f'Train: {train_mrr:.4f}, '
-                      f'Valid: {valid_mrr:.4f}, '
-                      f'Test: {test_mrr:.4f}')
+                if epoch % args.log_steps == 0:
+                    train_mrr, valid_mrr, test_mrr = result
+                    print(f'Run: {run + 1:02d}, '
+                          f'Epoch: {epoch:02d}, '
+                          f'Loss: {loss:.4f}, '
+                          f'Train: {train_mrr:.4f}, '
+                          f'Valid: {valid_mrr:.4f}, '
+                          f'Test: {test_mrr:.4f}')
 
         logger.print_statistics(run)
+
     logger.print_statistics()
