+++ examples/linkproppred/citation2/mf.py
@@ -105,7 +105,7 @@ def test(predictor, x, split_edge, evaluator, batch_size):
 
 
 def main():
-    parser = argparse.ArgumentParser(description='OGBL-Citation (MF)')
+    parser = argparse.ArgumentParser(description='OGBL-Citation2 (MF)')
     parser.add_argument('--device', type=int, default=0)
     parser.add_argument('--log_steps', type=int, default=1)
     parser.add_argument('--num_layers', type=int, default=3)
@@ -123,7 +123,7 @@ def main():
     device = f'cuda:{args.device}' if torch.cuda.is_available() else 'cpu'
     device = torch.device(device)
 
-    dataset = PygLinkPropPredDataset(name='ogbl-citation')
+    dataset = PygLinkPropPredDataset(name='ogbl-citation2')
     split_edge = dataset.get_edge_split()
     data = dataset[0]
 
@@ -140,7 +140,7 @@ def main():
     predictor = LinkPredictor(96, args.hidden_channels, 1, args.num_layers,
                               args.dropout).to(device)
 
-    evaluator = Evaluator(name='ogbl-citation')
+    evaluator = Evaluator(name='ogbl-citation2')
     logger = Logger(args.runs, args)
 
     for run in range(args.runs):
